DAOS-2203 object: EC update client changes.
Process full-stripe updates and add parity segments.
