+++ src/object/obj_internal.h
@@ -254,5 +254,10 @@ struct obj_ec_codec *obj_ec_codec_get(daos_oclass_id_t oc_id);
 int obj_encode_full_stripe(daos_obj_id_t oid, d_sg_list_t *sgl,
 			   uint32_t *sg_idx, size_t *sg_off,
 			   struct obj_ec_parity *parity, int p_idx);
-
+void
+ec_init_tgt_set(daos_iod_t *iods, unsigned int nr, struct daos_oclass_attr *oca,
+		unsigned long *tgt_set);
+int
+ec_obj_update(tse_task_t *task, daos_obj_id_t oid, daos_oclass_attr_t *oca,
+	      uint64_t *tgt_set);
 #endif /* __DAOS_OBJ_INTENRAL_H__ */
